Move files into containers
